"Remove user role restriction on file picker in Submission Page"
The "disabled" prop was removed from SingleFilePicker component on the Submission/Edit.js page, thereby removing the restriction that allowed only users with roles 8 and above to access the file picker. This change ensures all users, irrespective of their role field, have the ability to pick a file for submission.